requirements.txt
bcrypt==3.2.0
blinker==1.4
boto3==1.17.70
botocore==1.20.70
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dataclasses==0.6
dnspython==2.1.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
SQLAlchemy==1.4.9
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
